Remove client_secret dependency and add browser-auth flow
